+++ sklib/include/Integer.h
@@ -2,13 +2,14 @@
 #define INTEGER_H
 
 #include <Class.h>
+#include <stdint.h>
 
 typedef struct Integer {
-	int i;
+	int64_t i;
 } Integer;
 
 Class * integerClass();
 
-Integer * newInteger(int i);
+Integer * newInteger(int64_t i);
 
 #endif
